requirements.txt
Flask==3.0.3  #pip install Flask
Flask-SQLAlchemy==3.1.1  #pip install Flask-SQLAlchemy
Flask-Login==0.6.3  #pip install Flask-Login
Flask-Bcrypt==1.0.1  #pip install Flask-Bcrypt
python-dotenv==1.0.1  #pip install Flask-DotEnv
google-generativeai==0.8.3  #pip install -q -U google-generativeai
